Refactor the backward kernel for varlen reuse.
